fix(pr-mode): keep new comment on-screen and show detail loading state (Refs #20)
Two PR-mode usability fixes:

1. After posting a comment the detail view scrolled off-screen. The
   PrCommentCreated handler used a stale heuristic (~5 + body + comments)
   that ignored reviews, checks, section headers, and separators, so it
   under-scrolled whenever those sections were present. Route the post-create
   scroll through the canonical scroll_pr_detail_to_bottom, which derives the
   bottom from the pr_detail_content_line_count parity function, and delete the
   stale helpers. This is the same root cause class as the composer-open fix.

2. The PR detail pane appeared empty while detail data was being fetched, with
   no loading indication. Thread a detail_loading flag through
   build_pr_detail_content and pr_detail_content_line_count (kept in lockstep to
   preserve the scroll-clamp parity invariant) so the body renders a
   "Loading pull request..." indicator, and wire prs_state.loading.detail from
   the screen through the detail component.

Adds load-bearing regression coverage for the post-create scroll (with reviews
and checks populated) plus detail-loading render and parity tests. Splits the
PR render tests into pr_render_tests and pr_render_screen_tests to stay within
the per-file length budget.
